Move paddingParser to MJMLElement
